pyproject.toml
[tool.poetry]
name = "screen-scout"
version = "0.0.1"
description = "ScreenScout is a movie discovery platform that helps users explore and find their next favorite film with ease."
authors = ["sxw111"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {version = "^0.111.0", extras = ["all"]}
uvicorn = {version = "^0.30.5", extras = ["standart"]}
sqlalchemy = {version = "^2.0.31", extras = ["asyncio"]}
alembic = "^1.13.2"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
asyncpg = "^0.29.0"
asgi-lifespan = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
black = "^24.8.0"
mypy = "^1.11.1"
flake8 = "^7.1.1"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240327"
trio = "^0.26.1"
pytest-mock = "^3.14.0"
pre-commit = "^3.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
color=true
exclude = '''
    /(
        \.git
        | \.mypy_cache
        | \.venv
        | migrations/versions
    )/
'''
include = '\.pyi?$'
line-length = 88

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.isort]
profile = "black"
line_length = 88
